HER-41 move board as a subpage of a competition
