+++ app/competition/[name]/board/delegation/checking/page.tsx
