+++ app/competition/[name]/board/delegation/page.tsx
