+++ app/competition/[name]/board/draw/athlet/page.tsx
