+++ app/competition/[name]/board/draw/page.tsx
