+++ app/competition/[name]/board/match/bracket/page.tsx
