+++ app/competition/[name]/board/match/contest/page.tsx
