+++ app/competition/[name]/board/outline/page.tsx
