+++ app/competition/[name]/board/page.tsx
