+++ app/competition/[name]/board/results/page.tsx
