+++ app/competition/[name]/layout.tsx
