+++ app/competition/[name]/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const CompetitionDetailPage = ({ params }: { params: { name: string } }) => {
+  return <div>CompetitionDetailPage {params.name}</div>
+}
+
+export default CompetitionDetailPage
